Added extra input to transformation functions
Now, the third argument for char/text/back transformation functions will be the char/text/back located on the X/Y position of the buffer. If there is no such position on the window's buffer, it will be `nil`.
The `setCursorPos` function now floors inputs so as to not break when receiving non-integer arguments.